+++ Public/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### Release 0.0.2 Version
+
+#### Fixed
+-   **Dynamic Parameter Parsing**: Events can now be correctly published directly from chat with parameters that are automatically parsed based on the event's registered schema in Mods Communicator.
+-   **IEnumerable/ICollection/IList Support**: Collections are correctly constructed from space-separated values based on the registered type in event schema.
+
 ### Release 0.0.1 Version
 
 -   Initial release.